% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountCTR.R
\name{CountCTR}
\alias{CountCTR}
\title{Calculate and plot CTR}
\usage{
CountCTR(dt.path = NULL, metadata.path = NULL, dt = NULL, metadata = NULL)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Calculate correlation to reference datasets and plot related scatter plot
}
\examples{
CountCTR(dt=sample_data,metadata=sample_metadata)

}
