% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSNR.R
\name{CountSNR}
\alias{CountSNR}
\title{Calculate SNR
NOTE: Only PC1 and PC2}
\usage{
CountSNR(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Calculate SNR
NOTE: Only PC1 and PC2
}
\examples{
dt <- fread("./log2data.csv")
metadata <- fread("./metadata.csv")
CountSNR(dt,metadata)

}
