% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPerformance.R
\name{GetPerformance}
\alias{GetPerformance}
\title{Get output table}
\usage{
GetPerformance(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Get information of overall performance with historical data.
}
\examples{
GetPerformance(sample_data,sample_metadata)

}
