% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetReference.R
\docType{data}
\name{MetReference}
\alias{MetReference}
\title{Reference data of metabaolomics}
\format{
A \code{data.table} with 6 columns with 6 datasets combined.
}
\usage{
MetReference
}
\description{
A \code{data.table} file.
}
\keyword{datasets}
