% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_table.R
\name{PCA_table}
\alias{PCA_table}
\title{Plot PCA
NOTE: with all PCs}
\usage{
PCA_table(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{expDesign}{Data table, with a column named "col_names"}
}
\value{
Data table
}
\description{
Plot PCA
NOTE: with all PCs
}
\examples{
dt <- fread("./log2data.csv")
metadata <- fread("./metadata.csv")
PCA_table(dt,metadata)

}
