% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\docType{data}
\name{sample_metadata}
\alias{sample_metadata}
\title{The metadata of sample data for analysis}
\format{
A \code{data.table} with 6 columns. 
\describe{
\item{col_names}{A column needed for indentification of the ids of samples.}
}
}
\usage{
sample_metadata
}
\description{
A \code{data.table} file.
}
\keyword{datasets}
