% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeinRatios.R
\docType{data}
\name{SpikeinRatios}
\alias{SpikeinRatios}
\title{Ratios of spike-ins to Glycocholic acid (HMDB0000138)}
\format{
A \code{data.table} with 5 columns.
}
\usage{
SpikeinRatios
}
\description{
A \code{data.table} file.
}
\keyword{datasets}
