% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSNR.R
\name{CountSNR}
\alias{CountSNR}
\title{Calculate SNR and plot PCA}
\usage{
CountSNR(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Plot PCA and calculate SNR based on the first two principal componants of PCA.
}
\examples{
CountSNR(sample_data,sample_metadata)

}
