% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSNR.R
\name{CountSNR}
\alias{CountSNR}
\title{Calculate SNR}
\usage{
CountSNR(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Use the top two principal componants in PCA plot for the calculation of SNR.
}
\examples{
CountSNR(sample_data,sample_metadata)

}
