% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPerformance.R
\name{GetPerformance}
\alias{GetPerformance}
\title{Get output table}
\usage{
GetPerformance(
  dt.path = NULL,
  metadata.path = NULL,
  output.path = NULL,
  dt = NULL,
  metadata = NULL
)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Get information of overall performance with historical data.
}
\examples{
GetPerformance(dt=sample_data,metadata=sample_metadata)

}
