% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_table.R
\name{PCA_table}
\alias{PCA_table}
\title{Export table to plot PCA}
\usage{
PCA_table(dt, metadata)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table, with a column named "col_names"}
}
\value{
Data table
}
\description{
NOTE: with all PCs exported
}
\examples{
PCA_table(sample_data,sample_metadata)

}
